Move TextDifferencingServiceExtensions to a more applicable location
